% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{ddloglik}
\alias{ddloglik}
\title{Calculate double derivative of partial log likelihood (Hessian)}
\usage{
ddloglik(n, delta, z, beta)
}
\description{
Calculate double derivative of partial log likelihood (Hessian)
}
