% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{dloglik}
\alias{dloglik}
\title{Calculate derivative of partial log likelihood (score function)}
\usage{
dloglik(n, delta, z, beta)
}
\description{
Calculate derivative of partial log likelihood (score function)
}
